requirements.txt
aiogram
python-dotenv
spotipy
yt-dlp
requests
mutagen
